backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Google Cloud
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-documentai==2.21.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Caching and background tasks
redis==5.0.1
celery==5.3.4

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Validation and parsing
email-validator==2.1.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# File handling
aiofiles==23.2.1